# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a voda api server.",
        "title": "Voice Of Diary API (voda)",
        "contact": {
            "name": "API Support",
            "url": "https://minkj1992.github.io",
            "email": "minkj1992@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "exchange-diary-b4mzhzbzcq-du.a.run.app",
    "basePath": "/v1",
    "paths": {
...
        "/token": {
            "post": {
                "description": "AuthCode를 전달하여, access \u0026 refresh 토큰을 발급 받는다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "JWT 토큰 발급",
                "parameters": [
                    {
                        "description": "발급받은 AuthCode",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.tokenResponse"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/token/refresh": {
            "get": {
                "description": "refresh token을 전달하여, accessToken을 재발급받는다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "AccessToken 재발급",
                "parameters": [
                    {
                        "description": "refresh 토큰",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.TokenRefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.tokenResponse"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.TokenRefreshRequest": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "controller.TokenRequest": {
            "type": "object",
            "properties": {
                "authCode": {
                    "type": "string"
                },
                "deviceToken": {
                    "type": "string"
                }
            }
        },
...
        "controller.tokenResponse": {
            "type": "object",
            "properties": {
                "accesstoken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}